+++ exportplan/core/serializers.py
@@ -20,23 +20,10 @@ class CompanyObjectivesSerializer(serializers.Serializer):
     description = serializers.CharField(required=False, allow_blank=True, validators=[no_html])
     planned_reviews = serializers.CharField(required=False, allow_blank=True, validators=[no_html])
     owner = serializers.CharField(required=False, allow_blank=True, validators=[no_html])
-    start_month = serializers.IntegerField(required=False, allow_null=True, min_value=1, max_value=12)
-    start_year = serializers.IntegerField(required=False, allow_null=True, min_value=0, max_value=9999)
     end_month = serializers.IntegerField(required=False, allow_null=True, min_value=1, max_value=12)
     end_year = serializers.IntegerField(required=False, allow_null=True, min_value=0, max_value=9999)
     companyexportplan = serializers.IntegerField()
     pk = serializers.IntegerField()
-
-    def validate(self, data):
-        """
-        Check that start is not before finish.
-        """
-        if all([data.get('start_month'), data.get('start_year'), data.get('end_month'), data.get('end_year')]):
-            start = datetime.date(day=1, month=data['start_month'], year=data['start_year'])
-            end = datetime.date(day=1, month=data['end_month'], year=data['end_year'])
-            if start > end:
-                raise serializers.ValidationError('End date must occur after start date')
-        return data
 
 
 class AboutYourBuinessSerializer(serializers.Serializer):